requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.4
django-braces==1.14.0
django-crispy-forms==1.12.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==3.2
jwcrypto==1.0
numpy==1.21.0
oauthlib==3.1.1
pandas==1.3.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
utils==1.0.1
wrapt==1.12.1
